Rewriting backend menu to use MenuItem Fixing bug in NodeInterface/NodeTrait
