+++ administrator/components/com_menus/Helper/MenusHelper.php
@@ -12,11 +12,13 @@ namespace Joomla\Component\Menus\Administrator\Helper;
 defined('_JEXEC') or die;
 
 use Joomla\CMS\Factory;
+use Joomla\CMS\Filesystem\Folder;
 use Joomla\CMS\Helper\ContentHelper;
 use Joomla\CMS\Language\Associations;
 use Joomla\CMS\Language\Multilanguage;
 use Joomla\CMS\Language\Text;
 use Joomla\CMS\Menu\MenuHelper;
+use Joomla\CMS\Menu\MenuItem;
 use Joomla\CMS\Table\Table;
 use Joomla\Database\DatabaseInterface;
 use Joomla\Registry\Registry;
@@ -33,6 +35,15 @@ class MenusHelper extends ContentHelper
 	 * Defines the valid request variables for the reverse lookup.
 	 */
 	protected static $_filter = array('option', 'view', 'layout');
+
+	/**
+	 * List of preset include paths
+	 *
+	 * @var  array
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 */
+	protected static $presets = null;
 
 	/**
 	 * Gets a standard form of a link for lookups.
@@ -290,12 +301,13 @@ class MenusHelper extends ContentHelper
 	 * @param   boolean  $enabledOnly  Whether to load only enabled/published menu items.
 	 * @param   int[]    $exclude      The menu items to exclude from the list
 	 *
-	 * @return  array
+	 * @return  MenuItem  A root node with the menu items as children
 	 *
-	 * @since   3.8.0
+	 * @since   __DEPLOY_VERSION__
 	 */
 	public static function getMenuItems($menutype, $enabledOnly = false, $exclude = array())
 	{
+		$root = new MenuItem;
 		$db    = Factory::getContainer()->get(DatabaseInterface::class);
 		$query = $db->getQuery(true);
 
@@ -340,21 +352,45 @@ class MenusHelper extends ContentHelper
 
 		try
 		{
-			$menuItems = $db->loadObjectList();
+			$menuItems = $db->loadObjectList('id', '\Joomla\CMS\Menu\MenuItem');
 
-			foreach ($menuItems as &$menuitem)
+			foreach ($menuItems as $menuitem)
 			{
 				$menuitem->params = new Registry($menuitem->params);
+
+				// Resolve the alias item to get the original item
+				if ($menuitem->type == 'alias')
+				{
+					static::resolveAlias($menuitem);
+				}
+
+				if ($menuitem->link = in_array($menuitem->type, array('separator', 'heading', 'container')) ? '#' : trim($menuitem->link))
+				{
+					$menuitem->submenu    = array();
+					$menuitem->class      = $menuitem->img ?? '';
+					$menuitem->scope      = $menuitem->scope ?? null;
+					$menuitem->browserNav = $menuitem->browserNav ? '_blank' : '';
+				}
+
+				if ($menuitem->parent_id > 1)
+				{
+					if (isset($menuItems[$menuitem->parent_id]))
+					{
+						$menuItems[$menuitem->parent_id]->addChild($menuitem);
+					}
+				}
+				else
+				{
+					$root->addChild($menuitem);
+				}
 			}
 		}
 		catch (\RuntimeException $e)
 		{
-			$menuItems = array();
-
 			Factory::getApplication()->enqueueMessage(Text::_('JERROR_AN_ERROR_HAS_OCCURRED'), 'error');
 		}
 
-		return $menuItems;
+		return $root;
 	}
 
 	/**
@@ -367,7 +403,7 @@ class MenusHelper extends ContentHelper
 	 *
 	 * @throws  \Exception
 	 *
-	 * @since   3.8.0
+	 * @since   __DEPLOY_VERSION__
 	 */
 	public static function installPreset($preset, $menutype)
 	{
@@ -392,7 +428,7 @@ class MenusHelper extends ContentHelper
 	 *
 	 * @throws  \Exception
 	 *
-	 * @since   3.8.0
+	 * @since   __DEPLOY_VERSION__
 	 */
 	protected static function installPresetItems(&$items, $menutype, $parent = 1)
 	{
@@ -521,5 +557,338 @@ class MenusHelper extends ContentHelper
 				static::installPresetItems($item->submenu, $menutype, $item->id);
 			}
 		}
+	}
+
+	/**
+	 * Add a custom preset externally via plugin or any other means.
+	 * WARNING: Presets with same name will replace previously added preset *except* Joomla's default preset (joomla)
+	 *
+	 * @param   string  $name     The unique identifier for the preset.
+	 * @param   string  $title    The display label for the preset.
+	 * @param   string  $path     The path to the preset file.
+	 * @param   bool    $replace  Whether to replace the preset with the same name if any (except 'joomla').
+	 *
+	 * @return  void
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 */
+	public static function addPreset($name, $title, $path, $replace = true)
+	{
+		if (static::$presets === null)
+		{
+			static::getPresets();
+		}
+
+		if ($name == 'joomla')
+		{
+			$replace = false;
+		}
+
+		if (($replace || !array_key_exists($name, static::$presets)) && is_file($path))
+		{
+			$preset = new \stdClass;
+
+			$preset->name  = $name;
+			$preset->title = $title;
+			$preset->path  = $path;
+
+			static::$presets[$name] = $preset;
+		}
+	}
+
+	/**
+	 * Get a list of available presets.
+	 *
+	 * @return  \stdClass[]
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 */
+	public static function getPresets()
+	{
+		if (static::$presets === null)
+		{
+			// Important: 'null' will cause infinite recursion.
+			static::$presets = array();
+
+			static::addPreset('joomla', 'JLIB_MENUS_PRESET_JOOMLA', JPATH_ADMINISTRATOR . '/components/com_menus/presets/joomla.xml');
+			static::addPreset('modern', 'JLIB_MENUS_PRESET_MODERN', JPATH_ADMINISTRATOR . '/components/com_menus/presets/modern.xml');
+
+			// Load from template folder automatically
+			$app = Factory::getApplication();
+			$tpl = JPATH_THEMES . '/' . $app->getTemplate() . '/html/com_menus/presets';
+
+			if (is_dir($tpl))
+			{
+				$files = Folder::files($tpl, '\.xml$');
+
+				foreach ($files as $file)
+				{
+					$name  = substr($file, 0, -4);
+					$title = str_replace('-', ' ', $name);
+
+					static::addPreset(strtolower($name), ucwords($title), $tpl . '/' . $file);
+				}
+			}
+		}
+
+		return static::$presets;
+	}
+
+	/**
+	 * Load the menu items from a preset file into a hierarchical list of objects
+	 *
+	 * @param   string    $name      The preset name
+	 * @param   bool      $fallback  Fallback to default (joomla) preset if the specified one could not be loaded?
+	 * @param   MenuItem  $parent    Root node of the menu
+	 *
+	 * @return  MenuItem
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 */
+	public static function loadPreset($name, $fallback = true, $parent = null)
+	{
+		$presets = static::getPresets();
+
+		if (!$parent)
+		{
+			$parent = new MenuItem;
+		}
+
+		if (isset($presets[$name]) && ($xml = simplexml_load_file($presets[$name]->path, null, LIBXML_NOCDATA)) && $xml instanceof \SimpleXMLElement)
+		{
+			static::loadXml($xml, $parent);
+		}
+		elseif ($fallback && isset($presets['joomla']))
+		{
+			if (($xml = simplexml_load_file($presets['joomla']->path, null, LIBXML_NOCDATA)) && $xml instanceof \SimpleXMLElement)
+			{
+				static::loadXml($xml, $parent);
+			}
+		}
+
+		return $parent;
+	}
+
+	/**
+	 * Method to resolve the menu item alias type menu item
+	 *
+	 * @param   \stdClass  &$item  The alias object
+	 *
+	 * @return  void
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 */
+	public static function resolveAlias(&$item)
+	{
+		$obj = $item;
+
+		while ($obj->type == 'alias')
+		{
+			$params  = new Registry($obj->params);
+			$aliasTo = $params->get('aliasoptions');
+
+			$db = Factory::getDbo();
+			$query = $db->getQuery(true);
+			$query->select('a.id, a.link, a.type, e.element')
+				->from('#__menu a')
+				->where('a.id = ' . (int) $aliasTo)
+				->join('left', '#__extensions e ON e.id = a.component_id = e.id');
+
+			try
+			{
+				$obj = $db->setQuery($query)->loadObject();
+
+				if (!$obj)
+				{
+					$item->link = '';
+
+					return;
+				}
+			}
+			catch (\Exception $e)
+			{
+				$item->link = '';
+
+				return;
+			}
+		}
+
+		$item->id      = $obj->id;
+		$item->link    = $obj->link;
+		$item->type    = $obj->type;
+		$item->element = $obj->element;
+	}
+
+	/**
+	 * Parse the flat list of menu items and prepare the hierarchy of them using parent-child relationship.
+	 *
+	 * @param   MenuItem  $item  Menu item to preprocess
+	 *
+	 * @return  void
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 */
+	public static function preprocess($item)
+	{
+		// Resolve the alias item to get the original item
+		if ($item->type == 'alias')
+		{
+			static::resolveAlias($item);
+		}
+
+		if ($item->link = in_array($item->type, array('separator', 'heading', 'container')) ? '#' : trim($item->link))
+		{
+			$item->submenu    = array();
+			$item->class      = $item->img ?? '';
+			$item->scope      = $item->scope ?? null;
+			$item->browserNav = $item->browserNav ? '_blank' : '';
+		}
+	}
+
+	/**
+	 * Load a menu tree from an XML file
+	 *
+	 * @param   \SimpleXMLElement[]  $elements  The xml menuitem nodes
+	 * @param   MenuItem             $parent    The menu hierarchy list to be populated
+	 * @param   string[]             $replace   The substring replacements for iterator type items
+	 *
+	 * @return  void
+	 *
+	 * @since  __DEPLOY_VERSION__
+	 */
+	protected static function loadXml($elements, $parent, $replace = array())
+	{
+		foreach ($elements as $element)
+		{
+			if ($element->getName() != 'menuitem')
+			{
+				continue;
+			}
+
+			$select = (string) $element['sql_select'];
+			$from   = (string) $element['sql_from'];
+
+			/**
+			 * Following is a repeatable group based on simple database query. This requires sql_* attributes (sql_select and sql_from are required)
+			 * The values can be used like - "{sql:columnName}" in any attribute of repeated elements.
+			 * The repeated elements are place inside this xml node but they will be populated in the same level in the rendered menu
+			 */
+			if ($select && $from)
+			{
+				$hidden = $element['hidden'] == 'true';
+				$where  = (string) $element['sql_where'];
+				$order  = (string) $element['sql_order'];
+				$group  = (string) $element['sql_group'];
+				$lJoin  = (string) $element['sql_leftjoin'];
+				$iJoin  = (string) $element['sql_innerjoin'];
+
+				$db    = Factory::getDbo();
+				$query = $db->getQuery(true);
+				$query->select($select)->from($from);
+
+				if ($where)
+				{
+					$query->where($where);
+				}
+
+				if ($order)
+				{
+					$query->order($order);
+				}
+
+				if ($group)
+				{
+					$query->group($group);
+				}
+
+				if ($lJoin)
+				{
+					$query->leftJoin($lJoin);
+				}
+
+				if ($iJoin)
+				{
+					$query->innerJoin($iJoin);
+				}
+
+				$results = $db->setQuery($query)->loadObjectList();
+
+				// Skip the entire group if no items to iterate over.
+				if ($results)
+				{
+					// Show the repeatable group heading node only if not set as hidden.
+					if (!$hidden)
+					{
+						$child = static::parseXmlNode($element, $replace);
+						$parent->addChild($child);
+					}
+
+					// Iterate over the matching records, items goes in the same level (not $item->submenu) as this node.
+					foreach ($results as $result)
+					{
+						static::loadXml($element->menuitem, $parent, $result);
+					}
+				}
+			}
+			else
+			{
+				$item = static::parseXmlNode($element, $replace);
+
+				// Process the child nodes
+				static::loadXml($element->menuitem, $item, $replace);
+
+				$parent->addChild($item);
+			}
+		}
+	}
+
+	/**
+	 * Create a menu item node from an xml element
+	 *
+	 * @param   \SimpleXMLElement  $node     A menuitem element from preset xml
+	 * @param   string[]           $replace  The values to substitute in the title, link and element texts
+	 *
+	 * @return  \stdClass
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 */
+	protected static function parseXmlNode($node, $replace = array())
+	{
+		$item = new MenuItem;
+
+		$item->id         = null;
+		$item->type       = (string) $node['type'];
+		$item->title      = (string) $node['title'];
+		$item->link       = (string) $node['link'];
+		$item->element    = (string) $node['element'];
+		$item->class      = (string) $node['class'];
+		$item->icon       = (string) $node['icon'];
+		$item->browserNav = (string) $node['target'];
+		$item->access     = (int) $node['access'];
+		$item->setParams(new Registry(trim($node->params)));
+		$item->scope      = (string) $node['scope'] ?: 'default';
+		$item->submenu    = array();
+
+		if ($item->type == 'separator' && trim($item->title, '- '))
+		{
+			$item->getParams()->set('text_separator', 1);
+		}
+
+		if ($item->type == 'heading' || $item->type == 'container')
+		{
+			$item->link = '#';
+		}
+
+		// Translate attributes for iterator values
+		foreach ($replace as $var => $val)
+		{
+			$item->title   = str_replace("{sql:$var}", $val, $item->title);
+			$item->element = str_replace("{sql:$var}", $val, $item->element);
+			$item->link    = str_replace("{sql:$var}", $val, $item->link);
+			$item->class   = str_replace("{sql:$var}", $val, $item->class);
+			$item->icon    = str_replace("{sql:$var}", $val, $item->icon);
+		}
+
+		return $item;
 	}
 }
