+++ administrator/components/com_menus/View/Menu/XmlView.php
@@ -56,10 +56,10 @@ class XmlView extends BaseHtmlView
 
 		if ($menutype)
 		{
-			$items = MenusHelper::getMenuItems($menutype, true);
+			$root = MenusHelper::getMenuItems($menutype, true);
 		}
 
-		if (empty($items))
+		if ($root->hasChildren())
 		{
 			Log::add(Text::_('COM_MENUS_SELECT_MENU_FIRST_EXPORT'), Log::WARNING, 'jerror');
 
@@ -68,7 +68,7 @@ class XmlView extends BaseHtmlView
 			return;
 		}
 
-		$this->items = MenuHelper::createLevels($items);
+		$this->items = $root->getChildren();
 
 		$xml = new \SimpleXMLElement('<menu ' .
 			'xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" ' .
