+++ administrator/modules/mod_menu/Menu/CssMenu.php
@@ -14,9 +14,7 @@ defined('_JEXEC') or die;
 use Joomla\CMS\Application\CMSApplication;
 use Joomla\CMS\Component\ComponentHelper;
 use Joomla\CMS\Language\Text;
-use Joomla\CMS\Menu\MenuHelper;
-use Joomla\CMS\Menu\Node;
-use Joomla\CMS\Menu\Tree;
+use Joomla\CMS\Menu\MenuItem;
 use Joomla\CMS\Table\Table;
 use Joomla\CMS\Uri\Uri;
 use Joomla\Component\Menus\Administrator\Helper\MenusHelper;
@@ -31,13 +29,22 @@ use Joomla\Utilities\ArrayHelper;
 class CssMenu
 {
 	/**
-	 * The Menu tree object
+	 * The root of the menu
 	 *
-	 * @var    Tree
+	 * @var    MenuItem
 	 *
-	 * @since  3.8.0
+	 * @since  __DEPLOY_VERSION__
 	 */
-	protected $tree;
+	protected $root;
+
+	/**
+	 * An array of MenuItem nodes
+	 *
+	 * @var    MenuItem[]
+	 *
+	 * @since  __DEPLOY_VERSION__
+	 */
+	protected $nodes = [];
 
 	/**
 	 * The module options
@@ -66,6 +73,7 @@ class CssMenu
 	 */
 	protected $application;
 
+	protected $counter = 0;
 	/**
 	 * CssMenu constructor.
 	 *
@@ -76,23 +84,7 @@ class CssMenu
 	public function __construct(CMSApplication $application)
 	{
 		$this->application = $application;
-	}
-
-	/**
-	 * Get the current menu tree
-	 *
-	 * @return  Tree
-	 *
-	 * @since   3.8.0
-	 */
-	public function getTree()
-	{
-		if (!$this->tree)
-		{
-			$this->tree = new Tree;
-		}
-
-		return $this->tree;
+		$this->root = new MenuItem;
 	}
 
 	/**
@@ -107,7 +99,6 @@ class CssMenu
 	 */
 	public function load($params, $enabled)
 	{
-		$this->tree    = $this->getTree();
 		$this->params  = $params;
 		$this->enabled = $enabled;
 		$menutype      = $this->params->get('menutype', '*');
@@ -115,44 +106,44 @@ class CssMenu
 		if ($menutype === '*')
 		{
 			$name   = $this->params->get('preset', 'joomla');
-			$levels = MenuHelper::loadPreset($name);
+			$this->root = MenusHelper::loadPreset($name);
 		}
 		else
 		{
-			$items = MenusHelper::getMenuItems($menutype, true);
+			$this->root = MenusHelper::getMenuItems($menutype, true);
 
-			if ($this->enabled && $this->params->get('check', 1))
+			// Can we access everything important with this menu? Create a recovery menu!
+			if ($this->enabled
+				&& $this->params->get('check', 1)
+				&& $this->check($this->root, $this->params))
 			{
-				if ($this->check($items, $this->params))
-				{
-					$this->params->set('recovery', true);
+				$this->params->set('recovery', true);
 
-					// In recovery mode, load the preset inside a special root node.
-					$this->tree->addChild(new Node\Heading('MOD_MENU_RECOVERY_MENU_ROOT'), true);
+				// In recovery mode, load the preset inside a special root node.
+				$this->root = new MenuItem(['level' => 0]);
+				$heading = new MenuItem(['title' => 'MOD_MENU_RECOVERY_MENU_ROOT', 'type' => 'heading']);
+				$heading->level = $this->root->level + 1;
 
-					$levels = MenuHelper::loadPreset('joomla');
-					$levels = $this->preprocess($levels);
+				$heading = MenusHelper::loadPreset('joomla', true, $heading);
+				$this->root->addChild($heading);
 
-					$this->populateTree($levels);
+				$this->preprocess($this->root);
 
-					$this->tree->addChild(new Node\Separator);
+				$this->root->addChild(new MenuItem(['type' => 'separator']));
 
-					// Add link to exit recovery mode
-					$uri = clone Uri::getInstance();
-					$uri->setVar('recover_menu', 0);
+				// Add link to exit recovery mode
+				$uri = clone Uri::getInstance();
+				$uri->setVar('recover_menu', 0);
 
-					$this->tree->addChild(new Node\Url('MOD_MENU_RECOVERY_EXIT', $uri->toString()));
+				$this->root->addChild(new MenuItem(['title' => 'MOD_MENU_RECOVERY_EXIT', 'type' => 'url', 'url' => $uri->toString()]));
 
-					$this->tree->getParent();
-				}
+				return $this->root;
 			}
-
-			$levels = MenuHelper::createLevels($items);
 		}
 
-		$levels = $this->preprocess($levels);
+		$this->preprocess($this->root);
 
-		$this->populateTree($levels);
+		return $this->root;
 	}
 
 	/**
@@ -164,15 +155,15 @@ class CssMenu
 	 *
 	 * @since   3.8.0
 	 */
-	public function renderSubmenu($layoutFile)
+	public function renderSubmenu($layoutFile, $node)
 	{
 		if (is_file($layoutFile))
 		{
-			$children = $this->tree->getCurrent()->getChildren();
+			$children = $node->getChildren();
 
-			foreach ($children as $child)
+			foreach ($children as $current)
 			{
-				$this->tree->setCurrent($child);
+				$current->level = $node->level + 1;
 
 				// This sets the scope to this object for the layout file and also isolates other `include`s
 				require $layoutFile;
@@ -253,29 +244,31 @@ class CssMenu
 	/**
 	 * Filter and perform other preparatory tasks for loaded menu items based on access rights and module configurations for display
 	 *
-	 * @param   \stdClass[]  $items  The levelled array of menu item objects
+	 * @param   MenuItem  $item  A menu item to process
 	 *
 	 * @return  array
 	 *
 	 * @since   3.8.0
 	 */
-	protected function preprocess($items)
+	protected function preprocess($parent)
 	{
-		$result     = array();
 		$user       = $this->application->getIdentity();
 		$language   = $this->application->getLanguage();
 
 		$noSeparator = true;
+		$children = $parent->getChildren();
 
 		// Call preprocess for the menu items on plugins.
 		// Plugins should normally process the current level only unless their logic needs deep levels too.
-		$this->application->triggerEvent('onPreprocessMenuItems', array('com_menus.administrator.module', &$items, $this->params, $this->enabled));
+		$this->application->triggerEvent('onPreprocessMenuItems', array('com_menus.administrator.module', $children, $this->params, $this->enabled));
 
-		foreach ($items as $i => &$item)
+
+		foreach ($children as $item)
 		{
 			// Exclude item with menu item option set to exclude from menu modules
-			if ($item->params->get('menu_show', 1) == 0)
+			if ($item->getParams()->get('menu_show', 1) == 0)
 			{
+				$parent->removeChild($item);
 				continue;
 			}
 
@@ -285,6 +278,7 @@ class CssMenu
 			// Whether this scope can be displayed. Applies only to preset items. Db driven items should use un/published state.
 			if (($item->scope === 'help' && $this->params->get('showhelp', 1) == 0) || ($item->scope === 'edit' && !$this->params->get('shownew', 1)))
 			{
+				$parent->removeChild($item);
 				continue;
 			}
 
@@ -305,12 +299,14 @@ class CssMenu
 			// Exclude item if is not enabled
 			if ($item->element && !ComponentHelper::isEnabled($item->element))
 			{
+				$parent->removeChild($item);
 				continue;
 			}
 
 			// Exclude Mass Mail if disabled in global configuration
 			if ($item->scope === 'massmail' && ($this->application->get('massmailoff', 0) == 1))
 			{
+				$parent->removeChild($item);
 				continue;
 			}
 
@@ -336,6 +332,7 @@ class CssMenu
 
 				if (!$createFields)
 				{
+					$parent->removeChild($item);
 					continue;
 				}
 
@@ -355,6 +352,7 @@ class CssMenu
 
 				if (!$workflow)
 				{
+					$parent->removeChild($item);
 					continue;
 				}
 
@@ -362,6 +360,7 @@ class CssMenu
 			}
 			elseif ($item->element === 'com_config' && !$user->authorise('core.admin'))
 			{
+				$parent->removeChild($item);
 				continue;
 			}
 			elseif ($item->element === 'com_admin')
@@ -370,23 +369,29 @@ class CssMenu
 
 				if (isset($query['view']) && $query['view'] === 'sysinfo' && !$user->authorise('core.admin'))
 				{
+					$parent->removeChild($item);
 					continue;
 				}
 			}
 
 			if ($assetName && !$user->authorise(($item->scope === 'edit') ? 'core.create' : 'core.manage', $assetName))
 			{
+				$parent->removeChild($item);
 				continue;
 			}
 
 			// Exclude if link is invalid
 			if (!in_array($item->type, array('separator', 'heading', 'container')) && trim($item->link) === '')
 			{
+				$parent->removeChild($item);
 				continue;
 			}
 
 			// Process any children if exists
-			$item->submenu = $this->preprocess($item->submenu);
+			if ($item->hasChildren())
+			{
+				$this->preprocess($item);
+			}
 
 			// Populate automatic children for container items
 			if ($item->type === 'container')
@@ -394,20 +399,32 @@ class CssMenu
 				$exclude    = (array) $item->params->get('hideitems') ?: array();
 				$components = MenusHelper::getMenuItems('main', false, $exclude);
 
-				$item->components = MenuHelper::createLevels($components);
-				$item->components = $this->preprocess($item->components);
-				$item->components = ArrayHelper::sortObjects($item->components, 'text', 1, false, true);
+				// We are adding the nodes first to preprocess them, then sort them and add them again.
+				foreach ($components->getChildren() as $c)
+				{
+					$item->addChild($c);
+				}
+
+				$this->preprocess($item);
+				$children = ArrayHelper::sortObjects($item->getChildren(), 'text', 1, false, true);
+
+				foreach ($children as $c)
+				{
+					$item->addChild($c);
+				}
 			}
 
 			// Exclude if there are no child items under heading or container
-			if (in_array($item->type, array('heading', 'container')) && empty($item->submenu) && empty($item->components))
+			if (in_array($item->type, array('heading', 'container')) && !$item->hasChildren() && empty($item->components))
 			{
+				$parent->removeChild($item);
 				continue;
 			}
 
 			// Exclude help menu item if set such in mod_menu
 			if ($this->params->get('showhelp', 1) == 0 && $item->link == 'index.php?option=com_cpanel&view=help')
 			{
+				$parent->removeChild($item);
 				continue;
 			}
 
@@ -416,6 +433,7 @@ class CssMenu
 			{
 				if ($noSeparator)
 				{
+					$parent->removeChild($item);
 					continue;
 				}
 
@@ -439,91 +457,70 @@ class CssMenu
 			}
 
 			$item->text = Text::_($item->title);
-
-			$result[$i] = $item;
 		}
 
 		// If last one was a separator remove it too.
-		if ($noSeparator && isset($i))
-		{
-			unset($result[$i]);
-		}
+		$last = end($parent->getChildren());
 
-		return $result;
+		if ($last && $last->type === 'separator' && $last->getSibling(false) && $last->getSibling(false)->type === 'separator')
+		{
+			$parent->removeChild($last);
+		}
 	}
 
 	/**
-	 * Load the menu items from a hierarchical list of items into the menu tree
+	 * Method to get the CSS class name for an icon identifier or create one if
+	 * a custom image path is passed as the identifier
 	 *
-	 * @param   \stdClass[]  $levels  Menu items as a hierarchical list format
+	 * @param  MenuItem  $node
 	 *
-	 * @return  void
+	 * @return  string	CSS class name
 	 *
 	 * @since   3.8.0
 	 */
-	protected function populateTree($levels)
+	public function getIconClass($node)
 	{
-		foreach ($levels as $item)
+		$identifier = $node->class;
+
+		// Top level is special
+		if (trim($identifier) == '')
 		{
-			$class = $this->enabled ? $item->class : 'disabled';
-
-			if ($item->type === 'separator')
-			{
-				$this->tree->addChild(new Node\Separator($item->title));
-			}
-			elseif ($item->type === 'heading')
-			{
-				// We already excluded heading type menu item with no children.
-				$this->tree->addChild(new Node\Heading($item->title, $class, null, $item->icon), $this->enabled);
-
-				if ($this->enabled)
-				{
-					$this->populateTree($item->submenu);
-					$this->tree->getParent();
-				}
-			}
-			elseif ($item->type === 'url')
-			{
-				$cNode = new Node\Url($item->title, $item->link, $item->browserNav, $class, null, $item->icon);
-				$this->tree->addChild($cNode, $this->enabled);
-
-				if ($this->enabled)
-				{
-					$this->populateTree($item->submenu);
-					$this->tree->getParent();
-				}
-			}
-			elseif ($item->type === 'component')
-			{
-				$cNode = new Node\Component($item->title, $item->element, $item->link, $item->browserNav, $class, null, $item->icon);
-				$this->tree->addChild($cNode, $this->enabled);
-
-				if ($this->enabled)
-				{
-					$this->populateTree($item->submenu);
-					$this->tree->getParent();
-				}
-			}
-			elseif ($item->type === 'container')
-			{
-				// We already excluded container type menu item with no children.
-				$this->tree->addChild(new Node\Container($item->title, $item->class, null, $item->icon), $this->enabled);
-
-				if ($this->enabled)
-				{
-					$this->populateTree($item->submenu);
-
-					// Add a separator between dynamic menu items and components menu items
-					if (count($item->submenu) && count($item->components))
-					{
-						$this->tree->addChild(new Node\Separator);
-					}
-
-					$this->populateTree($item->components);
-
-					$this->tree->getParent();
-				}
-			}
+			return null;
 		}
+
+		// We were passed a class name
+		if (substr($identifier, 0, 6) == 'class:')
+		{
+			$class = substr($identifier, 6);
+		}
+		// We were passed background icon url. Build the CSS class for the icon
+		else
+		{
+			if ($identifier == null)
+			{
+				return null;
+			}
+
+			$class = preg_replace('#\.[^.]*$#', '', basename($identifier));
+			$class = preg_replace('#\.\.[^A-Za-z0-9\.\_\- ]#', '', $class);
+		}
+
+		$html = 'fa-fw fa fa-' . $class;
+
+		return $html;
+	}
+
+	/**
+	 * Create unique identifier
+	 *
+	 * @return  string
+	 *
+	 * @since   4.0.0
+	 */
+	public function getCounter()
+	{
+		$this->counter++;
+
+		return $this->counter;
 	}
 }
