--- libraries/src/Menu/Node/Container.php
@@ -1,21 +1,0 @@
-<?php
-/**
- * Joomla! Content Management System
- *
- * @copyright  Copyright (C) 2005 - 2018 Open Source Matters, Inc. All rights reserved.
- * @license    GNU General Public License version 2 or later; see LICENSE.txt
- */
-namespace Joomla\CMS\Menu\Node;
-
-defined('JPATH_PLATFORM') or die;
-
-/**
- * A Container type of node for MenuTree
- *
- * @see    Node
- *
- * @since  3.8.0
- */
-class Container extends Heading
-{
-}
